5 GB is a lot

+++ src/position.rs
@@ -1,11 +1,13 @@
-#[derive(Debug, PartialEq)]
+use serde::{Serialize, Deserialize};
+
+#[derive(Debug, PartialEq, Serialize, Deserialize)]
 pub struct Position {
-    pub x: f64,
-    pub y: f64,
+    pub x: i128,
+    pub y: i128,
 }
 
 impl Position {
-    pub fn new(x: f64, y: f64) -> Self {
+    pub fn new(x: i128, y: i128) -> Self {
         Self {
             x,
             y,
